requirements.txt
Flask
Flask-Sessions
requests
flask-sqlalchemy
